backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.11
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.0
python-dotenv==1.1.0
alembic==1.12.0
passlib==1.7.4
gunicorn==21.2.0
marshmallow==3.20.1
requests==2.31.0
